backends: Move SRv4 Power Board Legacy Backend
