+++ j5/backends/hardware/sr/v4/legacy/__init__.py
@@ -1,0 +1,5 @@
+"""Backends for Student Robotics v4 Legacy Firmware."""
+
+from .power_board import SRV4LegacyPowerBoardHardwareBackend
+
+__all__ = ["SRV4LegacyPowerBoardHardwareBackend"]
