+++ tests/backends/hardware/sr/v4/legacy/__init__.py
@@ -1,0 +1,1 @@
+"""Test SR v4 legacy hardware backends."""
